wip move pick methods in correct manager
